rewrite ThrotthledSignalSpec
