+++ src/test/scala/com/wire/signals/ThrottledSignalSpec.scala
@@ -21,77 +21,70 @@ import java.util.concurrent.atomic.AtomicReference
 
 import com.wire.signals.testutils._
 
-import org.scalatest._
 import org.threeten.bp.Instant
 
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
 import Threading.defaultContext
 import CancellableFuture.delayed
-import com.wire.signals.testutils.Uncontended.random
 
-class ThrottledSignalSpec extends FeatureSpec with Matchers with Inspectors with BeforeAndAfter {
+class ThrottledSignalSpec extends munit.FunSuite {
 
-  feature("Forwarding") {
+  test("throttle serial events") {
+    100 times spying { spy =>
+      val s = Signal(1)
+      val m = s.throttle(2.millis)
+      m(spy.capture)
+      assertEquals(spy.received.get.map(_._1), Vector[Int](1))
 
-    scenario("throttle serial events") {
-      100 times spying { spy =>
-        val s = Signal(1)
-        val m = s.throttle(2.millis)
-        m(spy.capture)
-        spy.received.get.map(_._1) should contain only 1
+      (2 to 3) foreach { v =>
+        Thread.sleep(1)
+        s ! v
+        s ! v + 10
+      }
 
-        (2 to 3) foreach { v =>
-          Thread.sleep(1)
-          s ! v
-          s ! v + 10
-        }
-
-        withDelay {
-          spy.received.get.last._1 shouldEqual 13
-        }
+      withDelay {
+        assertEquals(spy.received.get.last._1, 13)
       }
     }
-
-    scenario("throttle parallel events")(spying { spy =>
-      import FiniteDuration.FiniteDurationIsOrdered, spy._
-
-      received.set(Vector.empty[(Int, Instant)])
-      val s = Signal[Int]()
-      val m = s.throttle(50.millis)
-      m(capture)
-
-      val updates = Future.sequence((1 to 10000).map(n => delayed(random.nextInt(500).millis) {
-        s ! n
-        s ! n + 1000
-      }.future))
-
-      Await.result(updates, 5.seconds)
-      val sorted = received.get.map(_._2).sorted
-      val interval = sorted.zip(sorted.tail).map { case (aa, bb) => (aa until bb).toMillis.millis }
-
-      every(interval) should be >= 45.millis
-    })
   }
 
-  feature("Auto-wiring") {
-    scenario("wire and un-wire throttled signal") {
-      lazy val s = new IntSignal(0)
-      val m = s.throttle(100.millis)
-      s.isWired shouldBe false
+  test("throttle parallel events")(spying { spy =>
+    import FiniteDuration.FiniteDurationIsOrdered, spy._
 
-      val o = m { _ => () }
-      s.isWired shouldBe true
+    received.set(Vector.empty[(Int, Instant)])
+    val s = Signal[Int]()
+    val m = s.throttle(50.millis)
+    m(capture)
 
-      o.disable()
-      s.isWired shouldBe false
+    val updates = Future.sequence((1 to 10000).map(n => delayed(random.nextInt(500).millis) {
+      s ! n
+      s ! n + 1000
+    }.future))
 
-      o.enable()
-      s.isWired shouldBe true
+    Await.result(updates, 5.seconds)
+    val sorted = received.get.map(_._2).sorted
+    val interval = sorted.zip(sorted.tail).map { case (aa, bb) => (aa until bb).toMillis.millis }
 
-      o.destroy()
-      s.isWired shouldBe false
-    }
+    interval.foreach { time => assert(time >= 45.millis) }
+  })
+
+  test("wire and un-wire throttled signal") {
+    lazy val s = new IntSignal(0)
+    val m = s.throttle(100.millis)
+    assert(!s.isWired)
+
+    val o = m { _ => () }
+    assert(s.isWired)
+
+    o.disable()
+    assert(!s.isWired)
+
+    o.enable()
+    assert(s.isWired)
+
+    o.destroy()
+    assert(!s.isWired)
   }
 
   class Spy {
