+++ src/test/scala/com/wire/signals/testutils/package.scala
@@ -1,17 +1,36 @@
 package com.wire.signals
 
 import java.util.concurrent.atomic.AtomicReference
-
 import com.wire.signals.utils._
 import org.scalatest.Matchers.fail
 import org.threeten.bp.{Duration, Instant}
 
+import java.util.Random
 import scala.concurrent.duration.{FiniteDuration, _}
 import scala.annotation.tailrec
 import scala.concurrent.{Await, ExecutionContext, Future, TimeoutException}
 import scala.util.{Failure, Try}
 
 package object testutils {
+  sealed trait Roughly[V] {
+    type Tolerance
+    def roughlyEquals(other: V)(implicit tolerance: Tolerance): Boolean
+  }
+
+  object Roughly extends {
+    implicit class RoughlyInstant(val instant: Instant) extends Roughly[Instant] {
+      override type Tolerance = Long
+      override def roughlyEquals(other: Instant)(implicit tolerance: Long): Boolean =
+        this.instant.toEpochMilli >= other.toEpochMilli - tolerance &&
+          this.instant.toEpochMilli <= other.toEpochMilli + tolerance
+    }
+  }
+
+  private val localRandom = new ThreadLocal[Random] {
+    override def initialValue: Random = new Random
+  }
+
+  def random: Random = localRandom.get
 
   implicit class SignalToSink[A](val signal: Signal[A]) extends AnyVal {
     def sink: SignalSink[A] = returning(new SignalSink[A])(_.subscribe(signal)(EventContext.Global))
